@wulianglongrd 2023-02-27
I think benchmarks are harmless and help us evaluate the performance impact of this change.