@wulianglongrd 2023-02-27
In fact, these benchmarks are not written specifically for this PR, but to evaluate the performance of `SelectVirtualServices` and `selectServices` functions, which `convertIstioListenerToWrapper` calls.

By the way, `SelectVirtualServices` and `selectServices` internally use 2 layers of for loops for host matching. I have done some optimization tests, because the optimization effect is not universal, so those codes are not included in this PR.