@istio-testing 2023-02-27
@wulianglongrd: The following test **failed**, say `/retest` to rerun all failed tests or `/retest-required` to rerun all mandatory failed tests:

Test name | Commit | Details | Required | Rerun command
--- | --- | --- | --- | ---
integ-security-multicluster_istio | 8bf5d77e9d5f611827da55d3330656b83465d77b | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/43613/integ-security-multicluster_istio/1630208191369318400) | unknown | `/test integ-security-multicluster`

<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->